setup sq07_sh07
may also have done some damage to the Level Sequence for this shot (sq07_sh07).. I tried to hit "record" as an attempt to get a single frame out to use as 3D-layout, but it did a real number on the rest of the sequence.
I can't seem to get animation from the "animation asset" to appear in-camera or in a render anymore. The media plate seems to be the issue - if it is in front of the camera, when rendering, it acts as a massive sunglas and darkens everything.